Separated file for console has been added. Console can affect game via GameManager and PlayerManager
